Move button folder to /components
