+++ client/src/components/helpers/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Logo from '../small/logo/logo';
-import Anchor from '../small/buttons/anchor';
+import Anchor from '../buttons/anchor';
 
 // Add thousands divider
 export function numberFormatter(number: number): string {
