+++ client/src/components/helpers/sidebar/sortDefined.tsx
@@ -3,8 +3,8 @@ import { H5, Icon } from '@blueprintjs/core';
 import { BlueprintIcons_16Id } from '@blueprintjs/icons/lib/esm/generated/16px/blueprint-icons-16';
 
 import Sorts from '../../small/input/sorts';
-import PrimaryButton from '../../small/buttons/primaryButton';
-import SecondaryButton from '../../small/buttons/secondaryButton';
+import PrimaryButton from '../../buttons/primaryButton';
+import SecondaryButton from '../../buttons/secondaryButton';
 import css from '../../contexts/themes.scss';
 
 type sortDefinedProps = {
